Add strict permission mode to CHERI::check_pointer.
In the strict permission mode, which can be enabled by setting the
`EnforceStrictPermissions` templating argument to `true`,
`check_pointer` removes permissions which are not declared in
`Permissions` from the passed `ptr` reference.

`ptr` can now be either a reference to a pointer, or a pointer container
that supports `get` and `operator=`, e.g., a Capability or a smart
pointer.

This new mode is useful for detecting cases where compartments ask for
less permissions than they actually require.

Extend the test suite to test this new feature.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>