allocator-test: test races between ephemeral claims and claims.
This expands the current hazard tests to include a situation where:

T1: allocate `ptr` with heap quota 1
T2: ephemeral claim on `ptr`
T1: free `ptr` with heap quota 1
T1: claim `ptr` with heap quota 2
T2: release ephemeral claim on `ptr`

In such a case, quota 2 should have a claim on `ptr` and thus the
pointer should not be freed. However this test fails as the allocator
ignores quota 2's claim and frees `ptr`.

This is because there is no check for the number of claims in
`mspace_free`, which also assumes that all quotas have already been
refunded. It just proceeds painting and freeing once `ptr` leaves the
hazard list.

This corner case was reported by Reto Achermann (UBC) as part of an
ongoing verification effort of the memory allocator.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>